refactor completely after review
